refactor: Move components own folders
