+++ apps/web/src/components/dashboard/index.tsx
@@ -22,7 +22,7 @@ import EarnBanner, { earnBannerID } from '@/components/dashboard/NewsCarousel/ba
 import SpacesBanner, { spacesBannerID } from '@/components/dashboard/NewsCarousel/banners/SpacesBanner'
 import StakeBanner, { stakeBannerID } from '@/components/dashboard/NewsCarousel/banners/StakeBanner'
 import AddFundsToGetStarted from '@/components/dashboard/AddFundsBanner'
-import PositionsWidget from '@/features/positions/PositionsWidget'
+import PositionsWidget from '@/features/positions/components/PositionsWidget'
 
 const RecoveryHeader = dynamic(() => import('@/features/recovery/components/RecoveryHeader'))
