+++ apps/web/src/features/positions/components/PositionsWidget/index.tsx
@@ -17,8 +17,8 @@ import {
 import { ViewAllLink } from '@/components/dashboard/styled'
 import css from './styles.module.css'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import PositionsHeader from '@/features/positions/PositionsHeader'
-import Position from '@/features/positions/Position'
+import PositionsHeader from '@/features/positions/components/PositionsHeader'
+import Position from '@/features/positions/components/Position'
 import usePositions from '@/features/positions/hooks/usePositions'
 import DefiImage from '@/public/images/balances/defi.png'
 import Image from 'next/image'
